+++ Frends.FTP.UploadFiles/Frends.FTP.UploadFiles/Definitions/SingleFileTransferResult.cs
@@ -5,13 +5,9 @@ namespace Frends.FTP.UploadFiles.Definitions
     internal class SingleFileTransferResult
     {
         public bool Success { get; set; }
-
         public bool ActionSkipped { get; set; }
-
         public IList<string> ErrorMessages { get; set; }
-
         public string TransferredFile { get; set; }
-
         public string TransferredFilePath { get; set; }
 
         public SingleFileTransferResult()