+++ Frends.FTP.UploadFiles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.1] - 2022-11-25
+### Changed
+- Connection.BufferSize documentation update. Value is set as bytes instead of KBs.
+- Removed Connection.ClientCertificatePath's default value.
+- Updated Task's return fields documentation.
+
 ## [1.0.0] - 2022-05-29
 ### Changed
 - Initial implementation